chore: mv display detection code to IBaseWindow
