+++ src/vs/platform/windows/test/electron-main/windowsFinder.test.ts
@@ -55,7 +55,6 @@ suite('WindowsFinder', () => {
 			lastFocusTime = options.lastFocusTime;
 			isFullScreen = false;
 			isReady = true;
-			maximizedWindowState: IWindowState | undefined;
 
 			ready(): Promise<ICodeWindow> { throw new Error('Method not implemented.'); }
 			setReady(): void { throw new Error('Method not implemented.'); }
